+++ interfaces.go
@@ -26,7 +26,7 @@ type Database interface {
 	UpdateMetricsHeartbeat() error
 	GetPatterns() ([]string, error)
 	SubscribeMetricEvents(shutdown chan bool) <-chan *MetricEvent
-	GetMetricRetention(metric string) (int32, error)
+	GetMetricRetention(metric string) (int64, error)
 
 	SaveMetrics(buffer map[string]*MatchedMetric) error
 	GetMetricsValues(metrics []string, from int64, until int64) (map[string][]*MetricValue, error)
